\name{datos}
\alias{datos}
\docType{data}
\title{Potato regional trial in Colombia}
\description{
Eleven trials with ten genotypes stablished in state of Narino-Colombia
Cortesy: Professor Luis Ernesto Rodriguez Molano \email{lerodriguezmo@unal.edu.co} 
}
\usage{data("datos")}
\format{
  A data frame with 440 observations on the following 17 variables.
  \describe{
    \item{\code{Localidad}}{a numeric vector}
    \item{\code{Nlocalidad}}{a character vector}
    \item{\code{Semestre}}{a numeric vector}
    \item{\code{Bloque}}{a numeric vector}
    \item{\code{Genotipo}}{a character vector}
    \item{\code{RO}}{a numeric vector}
    \item{\code{R1}}{a numeric vector}
    \item{\code{Rcomercial}}{a numeric vector}
    \item{\code{RcTon}}{a numeric vector}
    \item{\code{R2}}{a numeric vector}
    \item{\code{R3}}{a numeric vector}
    \item{\code{RendimientoTotal}}{a numeric vector}
    \item{\code{RTton}}{a numeric vector}
    \item{\code{GE}}{a numeric vector}
    \item{\code{MS}}{a numeric vector}
    \item{\code{CF}}{a numeric vector}
    \item{\code{AR}}{a numeric vector}
  }
}
\references{
Unpublished data.
}
\examples{
data(datos)
}