% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_qstat.R
\name{ge_qstat}
\alias{ge_qstat}
\title{Return output from qstat command as data.frame}
\usage{
ge_qstat(cmd = "qstat", flags = "", user = NULL, queue = NULL,
  filter = NULL, exact = TRUE, ext = FALSE, raw = FALSE)
}
\arguments{
\item{cmd}{qstat command, default `qstat`}

\item{flags}{add flags}

\item{user}{specify user}

\item{queue}{specify queue}

\item{filter}{filter (afterwards) list of column names and filter values}

\item{exact}{filter using exact match or partial match? `TRUE` by default}
}
\description{
Return output from qstat command as data.frame
}

