% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_qsub.R
\name{ge_qsub}
\alias{ge_qsub}
\title{Submit job to GridEngine}
\usage{
ge_qsub(cmd, script_file = "job.sh", submit_cmd = "qsub", queue = NULL,
  name = NULL, prefix = NULL)
}
\arguments{
\item{cmd}{command to execute}

\item{script_file}{script file called by submit command}

\item{submit_cmd}{command used to submit SGE job (default `qsub`)}

\item{queue}{queue name}

\item{jobId}{job name. If blank assigned automatically by grid engine}
}
\description{
Submit job to GridEngine
}

