\name{m1}
\alias{m1}
\docType{data}
\title{Samples of the posterior distribution by GIBBS sampler}
\description{
Object obtained by function RGE
}
\usage{data("m1")}
\format{
  The format is:
 num [1:20, 1:10000] 14.2 15.8 16.1 19.9 17.3 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:20] "Predicted_T1" "Predicted_T2" "Predicted_UN 4" "Predicted_UN 50" ...
  ..$ : NULL
}
\examples{
data(m1)
}