% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_qdel.R
\name{qdel}
\alias{qdel}
\title{Kill jobs}
\usage{
qdel(id = NULL, force = FALSE, flags = c(), cmd = "qdel",
  verbose = TRUE)
}
\arguments{
\item{id}{job id to kill}

\item{force}{force kill (`FALSE` by default)}

\item{flags}{possible vector of other flags passed to qdel command}

\item{cmd}{command to use (default `qdel`)}

\item{verbose}{`T`/`F`}
}
\description{
Kill jobs
}

