% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_qhost.R
\name{qhost}
\alias{qhost}
\title{Return output from qhost command as data.frame}
\usage{
qhost(cmd = "qhost", flags = NULL, filter = NULL, exact = TRUE,
  ext = FALSE, raw = FALSE)
}
\arguments{
\item{cmd}{qstat command, default `qhost`}

\item{flags}{add flags}

\item{filter}{filter (afterwards) list of column names and filter values}

\item{exact}{filter using exact match or partial match? `TRUE` by default}

\item{user}{specify user}

\item{queue}{specify queue}
}
\description{
Return output from qhost command as data.frame
}

