\name{RGE-package}
\alias{RGE-package}
\alias{RGE}
\docType{package}
\title{\packageTitle{RGE}}
\description{
RGE is a packages for analize regionals trials from plant breeding programs. The package simplify the analysis process in order to obtaind the more useful results to be consider for the resercher. The program perfomace a GIBBS sampler and finally obtain a bayesian yield stability index.
Tools for obtain useful plot were developed in order to make the interpretation of results more easy.
}
\author{Jose Miguel Cotes Torres \email{jmcotes@unal.edu.co}}
\references{
Cotes Torres, J. M., Gonzalez Jaimes, E. P., & Cotes Torres, A. (2016). Seleccion de Genotipos con Alta Respuesta y Estabilidad Fenotipica en Pruebas Regionales: Recuperando el Concepto Biologico. Revista Facultad De Ciencias Basicas, 8(2), 226-243.}