% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_qstat.R
\name{sge_qstat}
\alias{sge_qstat}
\title{Return output from qstat command as data.frame}
\usage{
sge_qstat(cmd = "qstat", flags = "", user = NULL, filter = NULL,
  exact = TRUE)
}
\arguments{
\item{cmd}{qstat command, default `qstat`}

\item{flags}{add flags}

\item{user}{show only for specific user}

\item{filter}{filter list of column names and filter values}

\item{exact}{filter using exact match or partial match? `TRUE` by default}
}
\description{
Return output from qstat command as data.frame
}

