% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_qsub.R
\name{sge_qsub}
\alias{sge_qsub}
\title{Submit job to SGE}
\usage{
sge_qsub(cmd, script_file = "job.sh", submit_cmd = "qsub")
}
\arguments{
\item{cmd}{command to execute}

\item{script_file}{script file called by submit command}

\item{submit_cmd}{command used to submit SGE job (default `qsub`)}
}
\description{
Submit job to SGE
}

